Inline AbstractVariableWidthBlock
